pyproject.toml
[project]
name = "skill-repo"
version = "0.1.0"
description = "Skill 仓库 CLI 工具 - 管理和共享 code agent 技能"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click",
    "rich-click",
    "tomli",
    "tomli-w",
    "PyYAML",
    "questionary",
    "rich",
]

[project.scripts]
skill-repo = "skill_repo.cli:cli"

[dependency-groups]
dev = [
    "pytest",
    "hypothesis",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skill_repo"]

[tool.pytest.ini_options]
testpaths = ["tests"]
